+++ CHANGES.rst
@@ -1,9 +1,10 @@
 0.12 (unreleased)
 -----------------
 
-- Implement twilight sky brightness model.
+- Implement Twilight sky brightness model and integrate with Atmosphere model.
+- Remove obsolete sky_conditions parameter and only read tabulated dark sky.
 
-0.11 (2017-10-10)
+0.11 (2017-11-10)
 -----------------
 
 - Implement fast fiberloss calculator that interpolates GalSim results (#78).